feat: draft implementation of UniqueIpCalculator
The public interface consists of a constructor, a funcion to add an IP
address to the logs and a funciton to retreive the number of unique IP
addresses in the logs.

The implementation uses a channel to enqueue new IP addresses, and a
goroutine that adds them to a set.

NB: this first draft implementation contains a bug (tests do not pass)
due to concurrency.